refactor: 두 개의 컨트롤러 하나로 통합
